added register screen
