+++ Client/src/Pages/Credentials/RegisterUsers.css
@@ -1,0 +1,19 @@
+.Register-users {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    min-height: 100vh;
+    margin-top: 8vh;
+}
+
+.Register-users-form {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    width: 25ch;
+    margin: 30px;
+}
+
+. Register-users-paper {
+    width: fit-content;
+}
